Extract set_breakdown auto-population/resync into a LogSetBreakdown concern
