+++ app/models/log.rb
@@ -1,5 +1,6 @@
 class Log < ApplicationRecord
   include LogScoring
+  include LogSetBreakdown
 
   belongs_to :user, default: -> { Current.user }
   belongs_to :workout
@@ -13,7 +14,6 @@ class Log < ApplicationRecord
   validates :score_type, presence: true
 
   before_save :backfill_lifting_loads_from_score
-  before_validation :resync_trivial_set_breakdown
 
   def build_movement_logs
     workout.exercises_for_log_recording.each do |exercise|
@@ -75,41 +75,6 @@ class Log < ApplicationRecord
     end
     auto_populate_set_breakdown(movement_log, exercise)
     movement_log
-  end
-
-  # A single logged rep, or a dedicated single-lift weightlifting day, is unbroken by
-  # construction -- no self-report needed. reps: 0 is this app's existing "unspecified/max
-  # effort" sentinel (see ExercisePrescription#rep_prescription_metric), not a literal single
-  # rep, so it's excluded here and left unpopulated like any other not-yet-known case.
-  def auto_populate_set_breakdown(movement_log, exercise)
-    return if movement_log.reps.blank? || movement_log.reps.zero?
-    return unless movement_log.reps == 1 || single_weightlifting_exercise_day?(exercise)
-
-    movement_log.set_breakdown = [movement_log.reps]
-  end
-
-  def single_weightlifting_exercise_day?(exercise)
-    exercise.movement.family_weightlifting? &&
-      workout.exercises.one? &&
-      !exercise.reps_defined_by_interval?
-  end
-
-  def resync_trivial_set_breakdown
-    exercises = workout.exercises_for_log_recording
-    movement_logs.each_with_index do |movement_log, index|
-      exercise = exercises[index]
-      next unless resyncable_set_breakdown?(movement_log, exercise)
-
-      movement_log.set_breakdown = [movement_log.reps]
-    end
-  end
-
-  def resyncable_set_breakdown?(movement_log, exercise)
-    return false unless exercise
-
-    single_weightlifting_exercise_day?(exercise) &&
-      movement_log.reps.present? && movement_log.reps.positive? &&
-      movement_log.set_breakdown.compact.size <= 1
   end
 
   # Copies a selected prescription dimension onto the movement log's performance columns. reps and
